# design/api/openapi.yml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  version: '1.0'
  description: 'REST API to be used to create tableside food ordering applications.'
  contact:
    email: mawentowski@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4010/
    description: Prism proxy server
  - url: http://localhost:8080/
    description: API server
...
paths:
...
  /ingredients:
    post:
      summary: Create an ingredient
      description: |
        This operation creates a new ingredient with the provided information. Set an initial quantity of the ingredient using the `in_stock_qty` field.
            
        The `ingredient_id` returned in the response can be used to associate the ingredient with a dish. The `in_stock_qty` can be used to track if the ingredient has a positive quantity.
      operationId: postIngredient
      tags:
        - Ingredients
      security:
        - AccessToken: []
      requestBody:
        $ref: '#/components/requestBodies/CreateIngredientRequest'
      responses:
        '201':
          description: |
            Resource created successfully. The response body contains the newly created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
          headers:
            Location:
              schema:
                type: string
              description: Points to the resource endpoint for the newly created resource
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      summary: Get a list of ingredients
      description: This operation retrieves a list of ingredients based on the specified criteria. You can sort the results, include specific properties, select ingredients that match certain criteria, and paginate through the results.
      operationId: getIngredientList
      tags:
        - Ingredients
      parameters:
        - name: sort
          in: query
          description: |
            Sort fields in ascending or descending order. To sort ascending, postfix the field name with `:asc`. To sort descending, postfix the field name with `:desc`.
          schema:
            type: string
          example: 'in_stock_qty:asc'
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,in_stock_qty'
        - name: select
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator:value`.

            **Basic Equality (eq):**
            - For exact matches, use `eq:value`.

            **Comparison Operators:**
            - **Less Than (lt):** Use `lt:value`.
            - **Less Than or Equal To (lte):** Use `lte:value`.
            - **Greater Than (gt):** Use `gt:value`.
            - **Greater Than or Equal To (gte):** Use `gte:value`.

            Separate multiple filters with commas. Dates should be in `YYYY-MM-DD` format.
          schema:
            type: string
            example: 'in_stock_qty.eq:0'
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response. If `limit` is used in conjunction with `offset`, then page-based pagination is applied.
          schema:
            type: integer
            maximum: 30
          example: 20
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources. A value of `0` will display resources from the beginning of the list. Only affects the response data if a `limit` value is set.
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ingredient'
                  total_results:
                    type: integer
                    example: 1
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /ingredients/{ingredient_id}:
    get:
      summary: Get an ingredient's details
      description: This operation retrieves information about a specific ingredient identified by its unique ingredient ID.
      operationId: getIngredient
      tags:
        - Ingredients
      parameters:
        - name: ingredient_id
          in: path
          schema:
            type: string
          required: true
          example: '614d706b67c122001f35cc47'
          description: The unique identifier for the resource.
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'in_stock_qty,name'
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      summary: Update an ingredient
      description: This operation updates an ingredient using the provided information. Using this operation, API consumers can update the `in_stock_qty` of ingredients when orders are placed.
      operationId: putIngredient
      tags:
        - Ingredients
      security:
        - AccessToken: []
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: string
          example: '617d2e09ec42e0001f41bf84'
          description: The unique identifier for the resource.
      requestBody:
        $ref: '#/components/requestBodies/UpdateIngredientRequest'
      responses:
        '200':
          description: Successful operation. The response body contains the updated resource.
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      summary: Delete an ingredient
      description: This operation deletes an existing order identified by its unique order ID.
      operationId: deleteIngredient
      tags:
        - Ingredients
      security:
        - AccessToken: []
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: string
          example: '61c8c34bb02368001f8f0812'
          description: The unique identifier for the resource.
      responses:
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  requestBodies:
...
    CreateIngredientRequest:
      description: Request body containing information used to create the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                $ref: '#/components/schemas/Name'
              in_stock_qty:
                $ref: '#/components/schemas/InStockQty'
            required:
              - name
              - in_stock_qty
    UpdateIngredientRequest:
      description: Request body containing information used to update the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                $ref: '#/components/schemas/Name'
              in_stock_qty:
                $ref: '#/components/schemas/InStockQty'
...
  schemas:
...
    CreatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the resource was created in ISO 8601 format.
...
    Id:
      type: string
      example: '674d0bf5c28b69001f8e03a1'
      readOnly: true
      description: The unique identifier for the resource.
...
    Ingredient:
      type: object
      properties:
        ingredient_id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        name:
          $ref: '#/components/schemas/Name'
        in_stock_qty:
          $ref: '#/components/schemas/InStockQty'
      # required:
      #   - name
      #   - in_stock_qty
    InStockQty:
      type: integer
      description: The quantity of the ingredient in the restaurant's inventory. The quantity of the ingredient should be managed by the API consumer.
      example: 3
    Name:
      type: string
      example: 'main_course'
      description: ''
...
    Problem:
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
      required:
        - title
        - detail
...
    UpdatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the order was created in ISO 8601 format.
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.